requirements.txt
pdfplumber>=0.10.0
openpyxl>=3.1.0
hypothesis>=6.0.0
pytest>=7.0.0
Flask>=3.0.0
python-docx>=1.1.0
pdf2image>=1.16.0
Pillow>=10.0.0
PyMuPDF>=1.24.0
